requirements/base.txt
Django==3.1
djangorestframework==3.11.1
python_decouple==3.3
psycopg2-binary==2.8.5
djangorestframework-jwt==1.11.0
django-filter==2.3.0
Pillow==7.2.0
drf-yasg==1.17.1
sentry-sdk==0.16.1
requests==2.24.0
django-cors-headers==3.4.0